requirements.txt
asgiref==3.8.1
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-mapbox-location-field==2.1.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
